Mixin interface for flex layouts
