+++ src/main/java/com/vaadin/flow/component/orderedlayout/FlexConstants.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2000-2017 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.vaadin.flow.component.orderedlayout;
+
+class FlexConstants {
+
+    static final String JUSTIFY_CONTENT_CSS_PROPERTY = "justifyContent";
+    static final String FLEX_GROW_CSS_PROPERTY = "flexGrow";
+    static final String ALIGN_SELF_CSS_PROPERTY = "alignSelf";
+    static final String ALIGN_ITEMS_CSS_PROPERTY = "alignItems";
+
+}
