+++ src/test/java/com/vaadin/flow/component/orderedlayout/demo/HorizontalLayoutView.java
@@ -18,8 +18,8 @@ package com.vaadin.flow.component.orderedlayout.demo;
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.html.NativeButton;
-import com.vaadin.flow.component.orderedlayout.FlexLayout.Alignment;
-import com.vaadin.flow.component.orderedlayout.FlexLayout.JustifyContentMode;
+import com.vaadin.flow.component.orderedlayout.FlexComponent.Alignment;
+import com.vaadin.flow.component.orderedlayout.FlexComponent.JustifyContentMode;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.demo.DemoView;
 import com.vaadin.flow.router.Route;
